# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: JadeAI Career API
  version: 0.2.0
  description: |
    Living API contract for the JadeAI Career fork. Paths marked
    x-implementation-status: planned are not available yet.
servers:
  - url: /
security:
  - cookieSession: []
...
paths:
...
  /api/admin/users:
    get:
      tags: [Admin]
      operationId: listUsers
      x-requirement-id: AUTH-003
      x-implementation-status: implemented
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - name: query
          in: query
          schema:
            type: string
            maxLength: 120
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: Users
          content:
            application/json:
              schema:
                type: object
                required: [items, total, page, pageSize]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminUserSummary'
                  total:
                    type: integer
                    minimum: 0
                  page:
                    type: integer
                    minimum: 1
                  pageSize:
                    type: integer
                    minimum: 1
                    maximum: 100
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: jade_session
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
...
    Forbidden:
      description: Operation is not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message, requestId]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
...
    UserRole:
      type: string
      enum: [admin, user]
    UserStatus:
      type: string
      enum: [active, disabled, pending]
...
    AdminUserSummary:
      type: object
      additionalProperties: false
      required: [id, username, email, displayName, avatarUrl, authType, role, status, createdAt, lastLoginAt]
      properties:
        id:
          type: string
          format: uuid
        username:
          type: [string, 'null']
        email:
          type: [string, 'null']
          format: email
        displayName:
          type: string
        avatarUrl:
          type: [string, 'null']
          format: uri
        authType:
          type: string
          enum: [password, oauth, fingerprint]
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        createdAt:
          type: [string, 'null']
          format: date-time
        lastLoginAt:
          type: [string, 'null']
          format: date-time
...